pyproject.toml
[tool.poetry]
name = "finhan"
version = "1.0.0-alpha.0"
description = "Financial Projection for individuals and households"
authors = ["xtofl"]
license = "LGPL"
packages = [
    { include = "finhan", from ="src" }
]

[tool.poetry.scripts]
finhan-view = "finhan.main:main"
finhan-table = "finhan.adapter_table:main"
finhan-bepost-csvs = "finhan.adapter_bepost.csv_merge:main"
finhan-bepost-names = "finhan.adapter_bepost.csv_extract_account_names:main"

[tool.poetry.dependencies]
python = "^3.8"
matplotlib = "^3.3.4"
PyYAML = "^5.4.1"
coverage = "^5.5"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"

[tool.black]
line-length = 80

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
